# Improve performance and reduce allocations in MethodInjectionDirectiveBase