Improves performance and reduce allocations in **MethodInjectionDirectiveBase**.

**Before:**

|        Method |      Mean |    Error |   StdDev | Gen 0/1k Op | Gen 1/1k Op | Gen 2/1k Op | Allocated Memory/Op |
-------------- |----------:|---------:|---------:|------------:|------------:|------------:|--------------------:|
 ZeroArguments |  96.25 ns | 1.279 ns | 1.134 ns |      0.0304 |           - |           - |               128 B |
  TwoArguments | 335.92 ns | 5.264 ns | 4.666 ns |      0.2131 |           - |           - |               896 B |

**After:**

|        Method |      Mean |     Error |    StdDev | Gen 0/1k Op | Gen 1/1k Op | Gen 2/1k Op | Allocated Memory/Op |
-------------- |----------:|----------:|----------:|------------:|------------:|------------:|--------------------:|
 ZeroArguments |  16.29 ns | 0.3478 ns | 0.3083 ns |      0.0095 |           - |           - |                40 B |
  TwoArguments | 188.08 ns | 3.6042 ns | 3.1950 ns |      0.1810 |           - |           - |               760 B |